feat(nvim): update bash-language-server to 3.0.3
The default version of bash-language-server available in nixpkgs is old
and doesn't support shellcheck. This took me a long time to figure
out, but has a simple fix: upgrade to v3.0.3.

I followed a guide[1] to using node2nix with home-manager and generated
a package derivation for v3.0.3. Custom node packages can be updated by
running the following:

    cd nix/modules/nodejs && ./update.sh

Additional packages can be generated by adding them to the list in
node-packages.json.

[1] <https://code-notes.jhuizy.com/add-custom-npm-to-home-manager>