fix(nvim): Disable pop-up from sumneko_lua
The sumneko lua language server opens the following pop-up on startup:

> Do you need to configure your work environment as `luassert`?
>
> Request Actions:
> 1. Apply and modify settings
> 2. Apply but do not modify settings
> 3. Don't show again
> Type number and <Enter> or click with the mouse (q or empty cancels):

Disabling third party checks in the workspace config makes it stop. This
could be done in a .luarc.json, but that requires adding that file to
any lua project. Disabling it in the language server settings when
launched from nvim fixes it everywhere.