fix(nvim): use the nvim-treesitter specified parser versions
Currently nvim's healthcheck generates a ton of errors for treesitter
parsers due to using the parser versions from the treesitter upstream
instead of the versions locked by nvim-treesitter. This means a lot of
treesitter support ends up broken and I had to disable some languages
that I use to get around this.

Since I originally set this up nixpkgs has gained support[1] for using
the parser versions specified by nvim-treesitter, so we switch to that,
which fixes all of the healthcheck issues and means I can re-enable all
of the disabled languages.

[1]: https://nixos.wiki/wiki/Treesitter