chore: niv update
This update required a couple of fixes due to changes in nixpkgs.

First, NodeJS 18 is now EOL, so the custom nodePackages I was building
with node2nix now fail. I only set this up because nixpkgs had an old
version of bash-language-server, which has since been updated, so the
custom node packages mess can all just be removed.

Next, I have a custom vim module setup to pull in the vim-jjdescription
plugin which stopped building. Updating this to pull in the latest
revision from the Github repo fixes the build.