feat(git): Add script to generate commit message via Gemini
I've been playing around with generating commit messages via LLM to
provide a starting point. This adds a prepare-commit-msg hook script
that can be linked to from a repository. I've found it to be too slow
for much real use, but I'm keeping it around as a curiosity.

I did try to use a local model via ollama, but it was both slow and
generated crap messages. At least Gemini generates a good commit
message that tends to match up pretty closely to what I would write
manually.